refactor: polling service wires the message handler itself
- add non-generic BrokerageConcurrentMessageHandler with per-type Register, reusing the generic handler by composition
- restore BrokerageConcurrentMessageHandler<T> to master and drop BrokerageMessageQueue
- service constructor takes the handler, registers ProcessOrderState and routes the states; SeedAndStart drains the handler internally
- update the ADR wiring, seed-before-start and risk sections, and note the CharlesSchwab pilot
